EUW用スクリプトのリファクタリング
