+++ Source/PLATEAUEditorBPLibraries/Private/Export/PLATEAUExportModelBtn.cpp
@@ -1,6 +1,6 @@
 // Copyright © 2023 Ministry of Land, Infrastructure and Transport
 
-#include "PLATEAUExportModelBtn.h"
+#include "Export/PLATEAUExportModelBtn.h"
 #include "PLATEAUMeshExporter.h"
 
 /**
