+++ Source/PLATEAUEditorBPLibraries/Private/Import/PLATEAUImportModelBtn.cpp
@@ -1,6 +1,6 @@
 // Copyright © 2023 Ministry of Land, Infrastructure and Transport
 
-#include "PLATEAUImportModelBtn.h"
+#include "Import/PLATEAUImportModelBtn.h"
 #include "AssetSelection.h"
 #include "PLATEAUCityModelLoader.h"
 #include "PLATEAUEditor/Public/PLATEAUEditor.h"
