+++ Source/PLATEAUEditorBPLibraries/Private/ModelAdjustment/PLATEAUModelAdjustmentBuilding.cpp
@@ -1,6 +1,6 @@
 // Copyright © 2023 Ministry of Land, Infrastructure and Transport
 
-#include "PLATEAUModelAdjustmentBuilding.h"
+#include "ModelAdjustment/PLATEAUModelAdjustmentBuilding.h"
 #include "PLATEAUInstancedCityModel.h"
 using namespace citygml;
