+++ Source/PLATEAUEditorBPLibraries/Private/ModelAdjustment/PLATEAUModelAdjustmentVegetation.cpp
@@ -1,6 +1,6 @@
 // Copyright © 2023 Ministry of Land, Infrastructure and Transport
 
-#include "PLATEAUModelAdjustmentVegetation.h"
+#include "ModelAdjustment/PLATEAUModelAdjustmentVegetation.h"
 #include "PLATEAUImportSettings.h"
 using namespace citygml;
